+++ R/assert.R
@@ -11,13 +11,21 @@
 #'
 #' Argument to check.
 #'
-#' @param arg `[character]`
+#' @param ...
+#'
+#' Additional fields that will be added to the resulting error condition
+#'
+#' @param .arg `[character]`
 #'
 #' Name of argument being checked (used in error message).
 #'
-#' @param call `[call]`
+#' @param .call `[call]`
 #'
 #' Call to use in error message.
+#'
+#' @param .subclass `[character]`
+#'
+#' The (optional) subclass of the returned error condition.
 #'
 # -------------------------------------------------------------------------
 #' @return
@@ -38,7 +46,12 @@
 #'
 #' # Use in an internal function
 #' internal_fun <- function(a) {
-#'     .assert_string(a, arg = deparse(substitute(a)), call = sys.call(-1L))
+#'     .assert_string(
+#'         a,
+#'         .arg = deparse(substitute(x)),
+#'         .call = sys.call(-1L),
+#'         .subclass = "example_error"
+#'     )
 #'     TRUE
 #' }
 #' external_fun <- function(b) {
@@ -47,6 +60,7 @@
 #' external_fun(b="cat")
 #' try(external_fun())
 #' try(external_fun(b = letters))
+#' tryCatch(external_fun(b = letters), error = class)
 #'
 # -------------------------------------------------------------------------
 #' @name assertions
@@ -55,11 +69,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_integer <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!is.integer(x))
-        stopf("`%s` must be an integer vector.", arg, .call = call)
+.assert_integer <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!is.integer(x)) {
+        msg <- sprintf("`%s` must be an integer vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -71,11 +92,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_double <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!is.double(x))
-        stopf("`%s` must be a double vector.", arg, .call = call)
+.assert_double <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!is.double(x)) {
+        msg <- sprintf("`%s` must be a double vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -86,11 +114,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_numeric <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!is.numeric(x))
-        stopf("`%s` must be a numeric vector.", arg, .call = call)
+.assert_numeric <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!is.numeric(x)) {
+        msg <- sprintf("`%s` must be a numeric vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -101,11 +136,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_logical <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!is.logical(x))
-        stopf("`%s` must be a logical vector.", arg, .call = call)
+.assert_logical <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!is.logical(x)) {
+        msg <- sprintf("`%s` must be a logical vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -116,11 +158,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_character <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!is.character(x))
-        stopf("`%s` must be a character vector.", arg, .call = call)
+.assert_character <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!is.character(x)) {
+        msg <- sprintf("`%s` must be a character vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -131,31 +180,52 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_data_frame <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!is.data.frame(x))
-        stopf("`%s` must be a data frame.", arg, .call = call)
+.assert_data_frame <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!is.data.frame(x)) {
+        msg <- sprintf("`%s` must be a data frame.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_list <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!is.list(x))
-        stopf("`%s` must be a list.", arg, .call = call)
+.assert_list <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!is.list(x)) {
+        msg <- sprintf("`%s` must be a list.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_scalar_integer <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!(is.integer(x) && length(x) == 1L))
-        stopf("`%s` must be an integer vector of length 1.", arg, .call = call)
+.assert_scalar_integer <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!(is.integer(x) && length(x) == 1L)) {
+        msg <- sprintf("`%s` must be an integer vector of length 1.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -166,11 +236,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_scalar_integer_not_na <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!(is.integer(x) && length(x) == 1L) || is.na(x))
-        stopf("`%s` must be an integer vector of length 1 and not NA.", arg, .call = call)
+.assert_scalar_integer_not_na <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!(is.integer(x) && length(x) == 1L) || is.na(x)) {
+        msg <- sprintf("`%s` must be an integer vector of length 1 and not NA.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -181,11 +258,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_scalar_double <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!(is.double(x) && length(x) == 1L))
-        stopf("`%s` must be a double vector of length 1.", arg, .call = call)
+.assert_scalar_double <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!(is.double(x) && length(x) == 1L)) {
+        msg <- sprintf("`%s` must be a double vector of length 1.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -196,11 +280,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_scalar_double_not_na <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!(is.double(x) && length(x) == 1L) || is.na(x))
-        stopf("`%s` must be a double vector of length 1 and not NA.", arg, .call = call)
+.assert_scalar_double_not_na <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!(is.double(x) && length(x) == 1L) || is.na(x)) {
+        msg <- sprintf("`%s` must be a double vector of length 1 and not NA.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -211,11 +302,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_scalar_numeric <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!(is.numeric(x) && length(x) == 1L))
-        stopf("`%s` must be a numeric vector of length 1.", arg, .call = call)
+.assert_scalar_numeric <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!(is.numeric(x) && length(x) == 1L)) {
+        msg <- sprintf("`%s` must be a numeric vector of length 1.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -226,11 +324,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_scalar_numeric_not_na <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!(is.numeric(x) && length(x) == 1L) || is.na(x))
-        stopf("`%s` must be a numeric vector of length 1 and not NA.", arg, .call = call)
+.assert_scalar_numeric_not_na <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!(is.numeric(x) && length(x) == 1L) || is.na(x)) {
+        msg <- sprintf("`%s` must be a numeric vector of length 1 and not NA.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -241,11 +346,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_scalar_logical <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!(is.logical(x) && length(x) == 1L))
-        stopf("`%s` must be a logical vector of length 1.", arg, .call = call)
+.assert_scalar_logical <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!(is.logical(x) && length(x) == 1L)) {
+        msg <- sprintf("`%s` must be a logical vector of length 1.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -256,11 +368,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_bool <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!(is.logical(x) && length(x) == 1L) || is.na(x))
-        stopf("`%s` must be boolean (TRUE/FALSE).", arg, .call = call)
+.assert_bool <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!(is.logical(x) && length(x) == 1L) || is.na(x)) {
+        msg <- sprintf("`%s` must be boolean (TRUE/FALSE).", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -271,11 +390,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_scalar_character <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!(is.character(x) && length(x) == 1L))
-        stopf("`%s` must be a character vector of length 1.", arg, .call = call)
+.assert_scalar_character <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!(is.character(x) && length(x) == 1L)) {
+        msg <- sprintf("`%s` must be a character vector of length 1.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -286,12 +412,18 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_scalar_character_not_na <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!(is.character(x) && length(x) == 1L) || is.na(x))
-        stopf("`%s` must be a character vector of length 1 and not NA.", arg, .call = call)
-
+.assert_scalar_character_not_na <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+    if (!(is.character(x) && length(x) == 1L) || is.na(x)) {
+        msg <- sprintf("`%s` must be a character vector of length 1 and not NA.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -302,119 +434,193 @@ NULL
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_string <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_scalar_chr(x = x, arg = arg, call = call)
+.assert_string <- .assert_scalar_chr
+
+# -------------------------------------------------------------------------
+#' @rdname assertions
+#' @export
+.assert_non_negative_or_na <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+
+    if (!is.numeric(x)) {
+        msg <- sprintf("`%s` must be a numeric vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
+
+    if (!.all_non_negative_or_na(x)) {
+        msg <- sprintf("`%s` values must be non-negative or NA.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_non_negative_or_na <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
+.assert_non_positive_or_na <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
 
-    if (!is.numeric(x))
-        stopf("`%s` must be a numeric vector.", arg, .call = call)
+    if (!is.numeric(x)) {
+        msg <- sprintf("`%s` must be a numeric vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 
-    if (!.all_non_negative_or_na(x))
-        stopf("`%s` values must be non-negative or NA.", arg, .call = call)
-
+    if (!.all_non_positive_or_na(x)) {
+        msg <- sprintf("`%s` values must be non-positive or NA.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_non_positive_or_na <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
+.assert_non_negative <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
 
-    if (!is.numeric(x))
-        stopf("`%s` must be a numeric vector.", arg, .call = call)
+    if (!is.numeric(x)) {
+        msg <- sprintf("`%s` must be a numeric vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 
-    if (!.all_non_positive_or_na(x))
-        stopf("`%s` values must be non-positive or NA.", arg, .call = call)
-
-}
-
-# -------------------------------------------------------------------------
-#' @rdname assertions
-#' @export
-.assert_non_negative <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!is.numeric(x))
-        stopf("`%s` must be a numeric vector.", arg, .call = call)
-
-    if (!.all_non_negative(x))
-        stopf("`%s` values must be non-negative and not NA.", arg, .call = call)
-
-}
-
-# -------------------------------------------------------------------------
-#' @rdname assertions
-#' @export
-.assert_non_positive <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!is.numeric(x))
-        stopf("`%s` must be a numeric vector.", arg, .call = call)
-
-    if (!.all_non_positive(x))
-        stopf("`%s` values must be non-positive and not NA.", arg, .call = call)
-
-}
-
-# -------------------------------------------------------------------------
-#' @rdname assertions
-#' @export
-.assert_positive <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!is.numeric(x))
-        stopf("`%s` must be a numeric vector.", arg, .call = call)
-
-    if (!.all_positive(x))
-        stopf("`%s` values must be positive and not NA.", arg, .call = call)
-
-}
-
-# -------------------------------------------------------------------------
-#' @rdname assertions
-#' @export
-.assert_negative <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
-
-    if (!is.numeric(x))
-        stopf("`%s` must be a numeric vector.", arg, .call = call)
-
-    if (!.all_negative(x))
-        stopf("`%s` values must be negative and not NA.", arg, .call = call)
+    if (!.all_non_negative(x)) {
+        msg <- sprintf("`%s` values must be non-negative and not NA.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 
 }
 
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_positive_or_na <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
+.assert_non_positive <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
 
-    if (!is.numeric(x))
-        stopf("`%s` must be a numeric vector.", arg, .call = call)
+    if (!is.numeric(x)) {
+        msg <- sprintf("`%s` must be a numeric vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 
-    if (!.all_positive_or_na(x))
-        stopf("`%s` values must be positive or NA.", arg, .call = call)
 
+    if (!.all_non_positive(x)) {
+        msg <- sprintf("`%s` values must be non-positive and not NA.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
 #' @rdname assertions
 #' @export
-.assert_negative_or_na <- function(x, arg = deparse(substitute(x)), call = sys.call(-1L)) {
-    .assert_not_missing(x = x, arg = arg, call = call)
+.assert_positive <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
 
-    if (!is.numeric(x))
-        stopf("`%s` must be a numeric vector.", arg, .call = call)
+    if (!is.numeric(x)) {
+        msg <- sprintf("`%s` must be a numeric vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
 
-    if (!.all_negative_or_na(x))
-        stopf("`%s` values must be negative or NA.", arg, .call = call)
+    if (!.all_positive(x)) {
+        msg <- sprintf("`%s` values must be positive and not NA.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
+}
+
+# -------------------------------------------------------------------------
+#' @rdname assertions
+#' @export
+.assert_negative <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+
+    if (!is.numeric(x)) {
+        msg <- sprintf("`%s` must be a numeric vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
+
+    if (!.all_negative(x)) {
+        msg <- sprintf("`%s` values must be negative and not NA.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
+}
+
+# -------------------------------------------------------------------------
+#' @rdname assertions
+#' @export
+.assert_positive_or_na <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+
+    if (!is.numeric(x)) {
+        msg <- sprintf("`%s` must be a numeric vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
+
+    if (!.all_positive_or_na(x)) {
+        msg <- sprintf("`%s` values must be positive or NA.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
+}
+
+# -------------------------------------------------------------------------
+#' @rdname assertions
+#' @export
+.assert_negative_or_na <- function(
+    x,
+    ...,
+    .arg = deparse(substitute(x)),
+    .call = sys.call(-1L),
+    .subclass = NULL
+) {
+    .assert_not_missing(x, ..., .arg = .arg, .call = .call, .subclass = .subclass)
+
+    if (!is.numeric(x)) {
+        msg <- sprintf("`%s` must be a numeric vector.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
+
+    if (!.all_negative_or_na(x)) {
+        msg <- sprintf("`%s` values must be negative or NA.", .arg)
+        .stop(msg, ..., .call = .call, .subclass = .subclass)
+    }
+
 
 }
 
@@ -424,9 +630,11 @@ NULL
 # ------------------------------------------------------------------------- #
 # ------------------------------------------------------------------------- #
 
-.assert_not_missing <- function(x, arg, call) {
-    if (missing(x))
-        stopf("argument `%s` is missing, with no default.", arg, .call = call)
+.assert_not_missing <- function(x, ..., .arg, .call, .subclass) {
+    if (missing(x)) {
+        msg <- sprintf("argument `%s` is missing, with no default.", .arg)
+        .stop(msg, .call = .call, .subclass = .subclass)
+    }
 }
 
 # -------------------------------------------------------------------------
@@ -508,3 +716,8 @@ NULL
 }
 
 # -------------------------------------------------------------------------
+
+.stop <- function(msg, ..., .call = sys.call(-1L), .subclass = NULL) {
+    stop(errorCondition(msg, ..., class = .subclass, call = .call[1L]))
+}
+