+++ inst/tinytest/test_assertions.R
@@ -38,11 +38,11 @@ expect_fixed_error( fun(i=1L,  d=1,       l=NA,        chr="cat",   b=c(TRUE, TR
 
 # scalar assertions work nested_functions"
 internal_fun <- function(ii, dd, ll, chrchr, bb) {
-    .assert_scalar_int(ii, arg = deparse(substitute(ii)), call = sys.call(-1L))
-    .assert_scalar_dbl(dd, arg = deparse(substitute(dd)), call = sys.call(-1L))
-    .assert_scalar_lgl(ll, arg = deparse(substitute(ll)), call = sys.call(-1L))
-    .assert_string(chrchr, arg = deparse(substitute(chrchr)), call = sys.call(-1L))
-    .assert_bool(bb, arg = deparse(substitute(bb)), call = sys.call(-1L))
+    .assert_scalar_int(ii, .arg = deparse(substitute(ii)), .call = sys.call(-1L))
+    .assert_scalar_dbl(dd, .arg = deparse(substitute(dd)), .call = sys.call(-1L))
+    .assert_scalar_lgl(ll, .arg = deparse(substitute(ll)), .call = sys.call(-1L))
+    .assert_string(chrchr, .arg = deparse(substitute(chrchr)), .call = sys.call(-1L))
+    .assert_bool(bb, .arg = deparse(substitute(bb)), .call = sys.call(-1L))
     TRUE
 }
 
@@ -80,7 +80,7 @@ z <- NA_integer_
 expect_null(.assert_int(y))
 expect_null(.assert_scalar_int_not_na(y))
 expect_fixed_error(.assert_int(x), "`x` must be an integer vector.")
-expect_fixed_error(.assert_int(arg="TEST"), "argument `TEST` is missing, with no default.")
+expect_fixed_error(.assert_int(.arg="TEST"), "argument `TEST` is missing, with no default.")
 expect_fixed_error(.assert_scalar_int_not_na(z), "`z` must be an integer vector of length 1 and not NA.")
 
 
@@ -90,7 +90,7 @@ y <- 1L
 z <- NA_real_
 expect_null(.assert_dbl(x))
 expect_fixed_error(.assert_dbl(y), "`y` must be a double vector.")
-expect_fixed_error(.assert_dbl(arg="TEST"), "argument `TEST` is missing, with no default.")
+expect_fixed_error(.assert_dbl(.arg="TEST"), "argument `TEST` is missing, with no default.")
 expect_null(.assert_scalar_dbl_not_na(x))
 expect_fixed_error(.assert_scalar_dbl_not_na(z), "`z` must be a double vector of length 1 and not NA.")
 
@@ -100,7 +100,7 @@ y <- "cat"
 z <- NA_real_
 expect_null(.assert_num(x))
 expect_fixed_error(.assert_num(y), "`y` must be a numeric vector.")
-expect_fixed_error(.assert_num(arg="TEST"), "argument `TEST` is missing, with no default.")
+expect_fixed_error(.assert_num(.arg="TEST"), "argument `TEST` is missing, with no default.")
 expect_null(.assert_scalar_num_not_na(x))
 expect_fixed_error(.assert_scalar_num_not_na(z), "`z` must be a numeric vector of length 1 and not NA.")
 
@@ -110,7 +110,7 @@ y <- "cat"
 z <- NA_character_
 expect_null(.assert_chr(y))
 expect_fixed_error(.assert_chr(x), "`x` must be a character vector.")
-expect_fixed_error(.assert_chr(arg="TEST"), "argument `TEST` is missing, with no default.")
+expect_fixed_error(.assert_chr(.arg="TEST"), "argument `TEST` is missing, with no default.")
 expect_null(.assert_scalar_chr_not_na(y))
 expect_fixed_error(.assert_scalar_chr_not_na(z), "`z` must be a character vector of length 1 and not NA.")
 
@@ -119,7 +119,7 @@ x <- 1
 y <- NA
 expect_null(.assert_lgl(y))
 expect_fixed_error(.assert_lgl(x), "`x` must be a logical vector.")
-expect_fixed_error(.assert_lgl(arg="TEST"), "argument `TEST` is missing, with no default.")
+expect_fixed_error(.assert_lgl(.arg="TEST"), "argument `TEST` is missing, with no default.")
 
 # numeric assertions
 x <- 1
@@ -130,23 +130,23 @@ expect_null(.assert_scalar_num(x))
 expect_null(.assert_scalar_num(y))
 expect_fixed_error(.assert_scalar_num(z), "`z` must be a numeric vector of length 1.")
 expect_fixed_error(.assert_scalar_num(w), "`w` must be a numeric vector of length 1.")
-expect_fixed_error(.assert_scalar_num(arg="TEST"), "argument `TEST` is missing, with no default.")
+expect_fixed_error(.assert_scalar_num(.arg="TEST"), "argument `TEST` is missing, with no default.")
 expect_null(.assert_num(w))
 expect_fixed_error(.assert_num(z), "`z` must be a numeric vector.")
-expect_fixed_error(.assert_num(arg="TEST"), "argument `TEST` is missing, with no default.")
+expect_fixed_error(.assert_num(.arg="TEST"), "argument `TEST` is missing, with no default.")
 
 # data frame assertions
 l <- .subset(mtcars)
 expect_null(.assert_data_frame(mtcars))
 expect_fixed_error(.assert_data_frame(l), "`l` must be a data frame.")
-expect_fixed_error(.assert_data_frame(arg="TEST"), "argument `TEST` is missing, with no default.")
+expect_fixed_error(.assert_data_frame(.arg="TEST"), "argument `TEST` is missing, with no default.")
 
 # list assertions
 l <- list(1, b=2)
 b <- "bat"
 expect_null(.assert_list(l))
 expect_fixed_error(.assert_list(b), "`b` must be a list.")
-expect_fixed_error(.assert_list(arg="TEST"), "argument `TEST` is missing, with no default.")
+expect_fixed_error(.assert_list(.arg="TEST"), "argument `TEST` is missing, with no default.")
 
 # negativity/positivity assertions
 zero_length <- integer()